refactor(filter): optimize filter structure
