+++ packages/sheets-filter-ui/src/controllers/sheets-filter-ui-desktop.controller.ts
@@ -32,7 +32,7 @@ import { ChangeFilterByOperation, CloseFilterPanelOperation, FILTER_PANEL_OPENED
 import { SheetsFilterPanelService } from '../services/sheets-filter-panel.service';
 import { SmartToggleFilterShortcut } from './sheets-filter.shortcut';
 import { ClearFilterCriteriaMenuItemFactory, ReCalcFilterMenuItemFactory, SmartToggleFilterMenuItemFactory } from './sheets-filter.menu';
-import { SheetsFilterMobileUIController } from './sheets-filter-mobile-ui.controller';
+import { SheetsFilterUIMobileController } from './sheets-filter-ui-mobile.controller';
 
 export interface IUniverSheetsFilterUIConfig {
     menu: MenuConfig;
@@ -47,8 +47,8 @@ export const FILTER_PANEL_POPUP_KEY = 'FILTER_PANEL_POPUP';
 /**
  * This controller controls the UI of "filter" features. Menus, commands and filter panel etc. Except for the rendering.
  */
-@OnLifecycle(LifecycleStages.Rendered, SheetsFilterUIController)
-export class SheetsFilterUIController extends SheetsFilterMobileUIController {
+@OnLifecycle(LifecycleStages.Rendered, SheetsFilterUIDesktopController)
+export class SheetsFilterUIDesktopController extends SheetsFilterUIMobileController {
     constructor(
         private readonly _config: Partial<IUniverSheetsFilterUIConfig>,
         @Inject(Injector) private readonly _injector: Injector,
